// src/components/templates/layoutBlock/types.ts
import { ReactNode } from "react";

export type LayoutBlockType = {
  children?: ReactNode;
  xxsColumns1?: boolean;
  xxsColumns2?: boolean;
  xxsColumns3?: boolean;
  xxsColumns4?: boolean;
  xxsColumns5?: boolean;
  xxsColumns6?: boolean;
  xsColumns1?: boolean;
  xsColumns2?: boolean;
  xsColumns3?: boolean;
  xsColumns4?: boolean;
  xsColumns5?: boolean;
  xsColumns6?: boolean;
  smColumns1?: boolean;
  smColumns2?: boolean;
  smColumns3?: boolean;
  smColumns4?: boolean;
  smColumns5?: boolean;
  smColumns6?: boolean;
  mdColumns1?: boolean;
  mdColumns2?: boolean;
  mdColumns3?: boolean;
  mdColumns4?: boolean;
  mdColumns5?: boolean;
  mdColumns6?: boolean;
  lgColumns1?: boolean;
  lgColumns2?: boolean;
  lgColumns3?: boolean;
  lgColumns4?: boolean;
  lgColumns5?: boolean;
  lgColumns6?: boolean;
  xlColumns1?: boolean;
  xlColumns2?: boolean;
  xlColumns3?: boolean;
  xlColumns4?: boolean;
  xlColumns5?: boolean;
  xlColumns6?: boolean;
  maxHeight?: boolean;
  maxWidth?: boolean;
};